fix: import register color from utils
