+++ packages/theme/src/common/color-tokens/breadcrumb.ts
@@ -1,7 +1,6 @@
 import { localize } from '@opensumi/ide-core-common';
 
-import { registerColor } from '../color-registry';
-import { transparent, lighten, darken } from '../utils';
+import { transparent, registerColor, lighten, darken } from '../utils';
 
 import { foreground } from './base';
 import { editorBackground, editorWidgetBackground } from './editor';
