+++ packages/theme/src/common/color-tokens/checkbox.ts
@@ -1,6 +1,6 @@
 import { localize } from '@opensumi/ide-core-common';
 
-import { registerColor } from '../color-registry';
+import { registerColor } from '../utils';
 
 export const checkboxBorder = registerColor(
   'checkbox.border',
