+++ packages/theme/src/common/color-tokens/custom/checkbox.ts
@@ -1,6 +1,6 @@
 import { localize } from '@opensumi/ide-core-common';
 
-import { registerColor } from '../../color-registry';
+import { registerColor } from '../../utils';
 
 export const ktCheckboxHoverBackground = registerColor(
   'kt.checkbox.hoverBackground',
