+++ packages/theme/src/common/color-tokens/custom/input.ts
@@ -1,6 +1,7 @@
 import { localize } from '@opensumi/ide-core-common';
 
-import { registerColor, contrastBorder } from '../../color-registry';
+import { registerColor } from '../../utils';
+import { contrastBorder } from '../base';
 import { inputBorder, inputOptionActiveBorder } from '../input';
 
 export const ktInputBorder = registerColor(
