+++ packages/theme/src/common/color-tokens/menu.ts
@@ -1,6 +1,6 @@
 import { localize } from '@opensumi/ide-core-common';
 
-import { registerColor } from '../color-registry';
+import { registerColor } from '../utils';
 
 import { contrastBorder, activeContrastBorder, foreground } from './base';
 import { dropdownBackground, dropdownForeground } from './dropdown';
