+++ packages/theme/src/common/color-tokens/pick-view.ts
@@ -1,7 +1,7 @@
 import { localize } from '@opensumi/ide-core-common';
 
 import { Color } from '../color';
-import { registerColor } from '../color-registry';
+import { registerColor } from '../utils';
 
 import { activeContrastBorder, contrastBorder } from './base';
