+++ packages/theme/src/common/color-tokens/status-bar.ts
@@ -1,8 +1,7 @@
 import { localize } from '@opensumi/ide-core-common';
 
 import { Color } from '../../common/color';
-import { registerColor } from '../color-registry';
-import { darken } from '../utils';
+import { darken, registerColor } from '../utils';
 
 import { ACTIVITY_BAR_BADGE_BACKGROUND, ACTIVITY_BAR_BADGE_FOREGROUND } from './activity-bar';
 import { contrastBorder, errorForeground } from './base';
