+++ packages/theme/src/common/color-tokens/tab.ts
@@ -1,8 +1,7 @@
 import { localize } from '@opensumi/ide-core-common';
 
 import { Color } from '../../common/color';
-import { registerColor } from '../color-registry';
-import { transparent } from '../utils';
+import { transparent, registerColor } from '../utils';
 
 import { contrastBorder } from './base';
 import { hcActiveBorderColor, hcBorderColor } from './basic-color';
