+++ packages/theme/src/common/color-tokens/title-bar.ts
@@ -1,7 +1,6 @@
 import { localize } from '@opensumi/ide-core-common';
 
-import { registerColor } from '../color-registry';
-import { transparent } from '../utils';
+import { transparent, registerColor } from '../utils';
 
 import { contrastBorder } from './base';
